requirements.txt
# Core dependencies
pyyaml
hcloud
ansible-core
cryptography

# Development dependencies
pytest
pytest-asyncio
pytest-mock